Added Basic Environment & Improved Code Formatting
Added an environment using a placeholder asset (same size and shape as the final planned room asset) and adjusted the size and shape of the playable region to account for this, defined by  'PLAYABLE_REGION'.

Additionally changed the formatting of the code to improve readability, and allow for easier implementation of additional features later on.