Added Health, Levelling, Damage, Attacking and better Enemy Player Tracking
Added a level and xp system to the player that scales their stats (health and attack) with their level. Experience is gained from killing enemies which is now possible by attacking them (using the space bar). The damage of this attack scales with the player's attack stat.

Added invincibility frames to both the player and enemy in order to stop them dying too quickly.

Improved enemy tracking the player and moving towards them, additionally added an element of randomness to the enemy's movements, so that when multiple enemies are on screen they don't all move at the same time.