Added Game Over Menu & Added Comments
Added a game over screen that saves the players score and downloads it to a txt file.

Additionally, commented the code to improve readability, explain the purpose of certain parts and explain the thought process behind certain decisions.

Unless any unforeseen issues with the game are found in the near future, this project is now finished.